+++ src/components/PlanManagement/plan.vue
@@ -1,13 +1,13 @@
 <template>
   <div id="plan"
        v-loading="loading"
-       element-loading-text="加载中">
+       :element-loading-text="logadingText">
     <el-card class="box-card">
       <div slot="header" class="clearfix">
         <label>阀值名称:</label>
         <el-input style="width: 150px" v-model="thresholdName" placeholder="请输入阀值名称"></el-input>
         <el-button type="primary" @click="primary">查询</el-button>
-        <el-button type="success" @click="exportExcel(tableData,multipleSelection)">导出</el-button>
+        <el-button type="success" @click="exportExcel(tableData,exportMulti)">导出</el-button>
       </div>
       <!--<el-scrollbar-->
       <!--style="height: 100%;"-->
@@ -47,6 +47,28 @@
             :label="item.title"
             min-width="150"
             align="center">
+          </el-table-column>
+          <el-table-column
+            v-if="item.data==='ATTACHMENT'"
+            :label="item.title"
+            width="300"
+            align="center">
+            <template slot-scope="scope">
+              <template v-if="scope.row.ATTACHMENT" v-for="(item,index) in scope.row.ATTACHMENT">
+                <p v-if="item.type==='pdf'" style="text-align: right">
+                  <a  target="_blank" :href="item.PDF_FILE_DOWNLOAD_URL" class="attachment">{{index+'.'}}{{item.name}}</a>
+                  <i style="cursor:pointer" class="fa  fa-download" @click="downFile(item)" title="文件下载"></i>
+                </p>
+                <p v-if="item.type==='file'" style="text-align: right">
+                  <a  class="attachment" @click="preview(item)">{{index+'.'}}{{item.name}}</a>
+                  <i style="cursor:pointer" class="fa  fa-download" @click="downFile(item)" title="文件下载"></i>
+                </p>
+                <p v-if="item.type==='image'" style="text-align: right">
+                  <a  class="attachment" @click="preview(item)">{{index+'.'}}{{item.name}}</a>
+                  <i style="cursor:pointer" class="fa  fa-download" @click="downFile(item)" title="文件下载"></i>
+                </p>
+              </template>
+            </template>
           </el-table-column>
           <el-table-column
             v-if="item.data==='caozuo'"
@@ -131,6 +153,23 @@
               </el-upload>
             </el-form-item>
           </el-col>
+          <el-col :span="12">
+            <el-form-item label="已上传文件列表：">
+              <div class="upFile">
+                <ul style="list-style: none">
+                  <template v-for="(item,index) in upFileList">
+                    <li>
+                      <span style="float: left">{{item.name}} </span>
+                      <span style="float: right;" class="iconHover">
+                        <i class=" el-icon-circle-check "></i>
+                        <i class=" el-icon-close " @click="delFile(item,index)"></i>
+                      </span>
+                    </li>
+                  </template>
+                </ul>
+              </div>
+            </el-form-item>
+          </el-col>
           <el-col :span="24" style="text-align: right">
             <el-form-item style="margin-bottom: 0!important;">
               <el-button type="primary" @click="submitForm" :loading="loading">提交</el-button>
@@ -148,6 +187,7 @@
 </template>
 
 <script>
+  import {checkFileExt,download,downloadFile} from "../../utils/utils";
   export default {
     name: "plan",
     data(){
@@ -158,6 +198,8 @@
         currentPage4: 1,
         thresholdName:"",
         fileList: [],
+        upFileList:[],
+        exportMulti:[],
         upFileUrl:this.$url.uploadUrl,
         tableHeader:[
           {data:'',title:'',type:"selection"},
@@ -167,10 +209,12 @@
           {data:'ALARM_LEVEL',title:'报警级别',type:"normal"},
           {data:'ALARM_LEVEL_COLOR',title:'报警级别颜色',type:"normal"},
           {data:'COUNTERMEASURES',title:'应对措施',type:"normal"},
+          {data:'ATTACHMENT',title:'附件',},
           {data:'caozuo',title:'操作'},
         ],
         multipleSelection:[],
         loading: true,
+        logadingText:"加载中！",
         dialogVisible: false,
         ruleForm:{
           thresholdName:"",
@@ -202,6 +246,8 @@
         },
         bodyFalse:false,
         title:"",
+        previewTitle:"",
+        previewSrc:"",
         screenWidth:document.body.clientWidth,
         fileData:{
           bizId:"",
@@ -228,7 +274,61 @@
             if(res.status===200){
               setTimeout(()=>{
                 that.loading=false;
+                that.logadingText="加载中！";
+                that.fileList=[];
                 let data=res.data.result;
+                $.each(data,(v,item)=>{
+                  // item.CREATE_TIME=new Date(item.CREATE_TIME).formatDate('yyyy-MM-dd');
+                  if (item.ATTACHMENT) {
+                    item.ATTACHMENT = JSON.parse(item.ATTACHMENT);
+                    $.each(item.ATTACHMENT,(s,items)=>{
+                      switch (items.type){
+                        case "jpg":
+                          items.type="image";
+                          break;
+                        case  "gif":
+                          items.type="image";
+                          break;
+                        case  "png":
+                          items.type="image";
+                          break;
+                        case  "jpeg":
+                          items.type="image";
+                          break;
+                        case  "zip":
+                          items.type="file";
+                          break;
+                        case  "rar":
+                          items.type="file";
+                          break;
+                        case  "xls":
+                          items.type="file";
+                          break;
+                        case  "xlsx":
+                          items.type="file";
+                          break;
+                        case  "doc":
+                          items.type="pdf";
+                          break;
+                        case  "docx":
+                          items.type="pdf";
+                          break;
+                        case  "pdf":
+                          items.type="pdf";
+                          break;
+                        case  "ppt":
+                          items.type="pdf";
+                          break;
+                        case  "pptx":
+                          items.type="pdf";
+                          break;
+                        default:
+                          items.type="";
+                          break
+                      }
+                    });
+                  };
+                });
                 that.tableData=data;
               },500);
 
@@ -256,11 +356,14 @@
       handleSelectionChange(val) {
         if(val.length>0){
           this.multipleSelection=[];
+          this.exportMulti=[];
           $.each(val,(v,item)=>{
             this.multipleSelection.push(item.PLAN_ID);
+            this.exportMulti.push(item);
           });
         }else{
           this.multipleSelection=[];
+          this.exportMulti=[];
         }
       },
       //点击行选中
@@ -315,6 +418,7 @@
         this.ruleForm.alarmLevelColor=row.ALARM_LEVEL_COLOR;
         this.ruleForm.countermeasures=row.COUNTERMEASURES;
         this.ruleForm.planId=row.PLAN_ID;
+        this.upFileList=row.ATTACHMENT;
       },
       //行内删除
       handleDelete(index, row) {
@@ -325,21 +429,14 @@
       //重置
       resetForm() {
         this.$refs['ruleForm'].resetFields();
-        this.fileList=[];
-        this.ruleForm={
-          thresholdName:"",
-          thresholdValue:"",
-          thresholdValueType:"",
-          alarmLevel:"",
-          alarmLevelColor:"red",
-          countermeasures:"",
-        };
         this.dialogVisible=false;
       },
       dialogClose(ruleForm){
         this.dialogVisible=false;
         this.$refs[ruleForm].resetFields();
-        this.fileList=[];
+        this.$refs.multipleTable.clearSelection();
+        this.upFileList=[];
+        this.exportMulti=[];
         this.ruleForm={
           thresholdName:"",
           thresholdValue:"",
@@ -350,21 +447,29 @@
         };
       },
       del(){
-        this.$http.delete(this.$url.baseUrl+'api/guoYang/auxiliary-decision/v0.1/gy-fxkh-plan-manage/delete',{data:this.multipleSelection}).then((res)=>{
-          if(res.status===200){
-            this.$message({
-              type:"success",
-              message:"删除成功！"
-            });
-            this.loading=true;
-            this.search();
-          }else{
-            this.$message({
-              type:"error",
-              message:"删除失败！"
-            })
-          }
-        });
+        if(this.multipleSelection.length>0){
+          this.$http.delete(this.$url.baseUrl+'api/guoYang/auxiliary-decision/v0.1/gy-fxkh-plan-manage/delete',{data:this.multipleSelection}).then((res)=>{
+            if(res.status===200){
+              this.$message({
+                type:"success",
+                message:"删除成功！"
+              });
+              this.loading=true;
+              this.search();
+            }else{
+              this.$message({
+                type:"error",
+                message:"删除失败！"
+              })
+            }
+          });
+        }else{
+          this.$message({
+            type:"error",
+            message:"请选择需要删除的数据！"
+          })
+        }
+
       },
       add(){
         this.dialogVisible=true;
@@ -389,45 +494,27 @@
             }
             _this.$http.put(url,_this.ruleForm).then((res)=>{
               if(res.status===200){
+                _this.dialogVisible=false;
+                _this.loading=true;
                 if(_this.fileList.length>0){
-                  _this.loading=true;
                   _this.fileData.bizId=res.data.result.message;
                   _this.$refs.upload.submit();
+                  _this.logadingText="文件上传中！"
                 }else{
-                  _this.$message({
-                    type:"success",
-                    message:msg
-                  });
-                  _this.dialogVisible=false;
-                  _this.loading=true;
-                  _this.multipleSelection=[];
-                  _this.fileList=[];
-                  _this.search();
-                  _this.$refs['ruleForm'].resetFields();
-                  this.ruleForm={
-                    thresholdName:"",
-                    thresholdValue:"",
-                    thresholdValueType:"",
-                    alarmLevel:"",
-                    alarmLevelColor:"red",
-                    countermeasures:"",
-                  };
+                  this.search();
                 }
+                _this.$message({
+                  type:"success",
+                  message:msg
+                });
+                _this.$refs['ruleForm'].resetFields();
               }else{
                 _this.$message({
                   type:"error",
                   message:msg1
                 });
+                _this.dialogVisible=false;
                 _this.$refs['ruleForm'].resetFields();
-                _this.fileList=[];
-                this.ruleForm={
-                  thresholdName:"",
-                  thresholdValue:"",
-                  thresholdValueType:"",
-                  alarmLevel:"",
-                  alarmLevelColor:"red",
-                  countermeasures:"",
-                };
               }
             })
           } else {
@@ -448,6 +535,16 @@
        * @param file
        */
       handleChange(file, fileList) {
+        $.each(fileList,(v,item)=>{
+          let fileExt = item.name.substring(item.name.lastIndexOf("."));
+          if (!checkFileExt(fileExt)) {
+            this.$message({
+              type:"error",
+              message:item.name+"的文件格式不符，请重新选择(格式:.jpg;.png;.jpeg;.gif;.zip;.rar;.doc;.docx;.xls;.xlsx;.pdf;.ppt;.pptx;)"
+            });
+            fileList.splice(v,1);
+          }
+        });
         this.fileList=fileList;
       },
       /**
@@ -463,24 +560,48 @@
         return  isLt10M;
       },
       handleAvatarSuccess(res, file){
-        if(res){
-          this.dialogVisible=false;
-          this.loading=true;
+        setTimeout(()=>{
+          this.$refs.upload.clearFiles();
+          this.search();
           this.$message({
             type:"success",
             message:"添加文件成功"
           });
-          this.search();
-        }else{
-          this.dialogVisible=false;
-          this.loading=true;
-          this.$message({
-            type:"error",
-            message:"添加文件失败"
-          });
-          this.search();
+        },500);
+      },
+      /**
+       * 删除文件
+       */
+      delFile(item){
+        this.$http.delete(this.$url.baseUrl+'api/attachment/v0.1/attachment',{data:{ids:[item.id]}}).then((res)=>{
+          if(res.status===200){
+            this.$message({
+              type:'success',
+              message:"删除成功"
+            });
+            this.upFileList.splice(item,1)
+          }
+        })
+      },
+      /**
+       * 文件预览
+       */
+      preview(item){
+        if(item.type==='file'){
+          download(item.url,'get');
+        }else if(item.type==="image"){
+          this.previewSrc=item.url;
+          this.previewTitle=item.name;
+          this.previewVisible=true;
         }
-      }
+      },
+      /**
+       * 下载文件
+       * @param item
+       */
+      downFile(item){
+        downloadFile(this.$url.baseUrl+'api/attachment/v0.1/attachment/download',[item.id])
+      },
     },
     computed:{
       tables(){
@@ -545,8 +666,19 @@
   #plan .table-button .add:hover{
     color: #0a95ef;
   }
+  #plan .attachment{
+    text-decoration: underline;
+    float: left;
+    overflow: hidden;
+    text-overflow:ellipsis;
+    white-space: nowrap;
+    cursor: pointer;
+    color: #0a95ef;
+    display: inline-block;
+    /*width: 100%;*/
+  }
 </style>
-<style>
+<style lang="less">
   #plan .table-button .add span{
     margin-left: 5px!important;
   }
@@ -557,5 +689,40 @@
     padding: 20px!important;
     height: 350px;
     overflow-y: auto;
+  }
+  #plan .upFile{
+    height: 150px;
+    overflow-y: auto;
+    ul{
+      li{
+        line-height: 32px;
+        width: 100%;
+        display: inline-block;
+        .iconHover{
+          .el-icon-circle-check{
+            display: block;
+            color:#67c23a ;
+            line-height: 32px;
+            margin-right: 10px;
+          }
+          .el-icon-close{
+            display: none;
+            color:#666666;
+            cursor: pointer;
+            line-height: 32px;
+            margin-right: 10px;
+          }
+        }
+
+        .iconHover:hover{
+          .el-icon-circle-check{
+            display: none;
+          }
+          .el-icon-close{
+            display: block;
+          }
+        }
+      }
+    }
   }
 </style>
