+++ src/components/floodControl/rescue.vue
@@ -23,7 +23,7 @@
             </el-option>
           </el-select>
           <el-button type="primary" @click="primary">查询</el-button>
-          <el-button type="success" @click="exportExcel(tableData,multipleSelection)">导出</el-button>
+          <el-button type="success" @click="exportExcel(tableData,exportMulti)">导出</el-button>
         </div>
         <!--<el-scrollbar-->
         <!--style="height: 100%;"-->
@@ -161,6 +161,7 @@
           {data:'caozuo',title:'操作'},
         ],
         multipleSelection:[],
+        exportMulti:[],
         loading: true,
         dialogVisible: false,
         cOptions:[],
@@ -257,11 +258,14 @@
       handleSelectionChange(val) {
         if(val.length>0){
           this.multipleSelection=[];
+          this.exportMulti=[];
           $.each(val,(v,item)=>{
             this.multipleSelection.push(item.RT_ID);
+            this.exportMulti.push(item);
           });
         }else{
           this.multipleSelection=[];
+          this.exportMulti=[];
         }
       },
       //点击行选中
@@ -333,17 +337,13 @@
       //重置
       resetForm() {
         this.$refs['ruleForm'].resetFields();
-        this.ruleForm={
-          rtName:"",
-          rtType:'常驻抢险队伍',
-          teamMemberInfo:"",
-          userId:"",
-        };
         this.dialogVisible=false;
       },
       dialogClose(ruleForm){
         this.dialogVisible=false;
         this.$refs[ruleForm].resetFields();
+        this.$refs.multipleTable.clearSelection();
+        this.exportMulti=[];
         this.ruleForm={
           rtName:"",
           rtType:'常驻抢险队伍',
@@ -358,21 +358,30 @@
        * 抢险队伍删除
        */
       del(){
-        this.$http.delete(this.$url.baseUrl+'api/guoYang/rescue-team/v0.1/team',{data:this.multipleSelection}).then((res)=>{
-          if(res.status===200){
-            this.$message({
-              type:"success",
-              message:"删除成功！"
-            });
-            this.loading=true;
-            this.search();
-          }else{
-            this.$message({
-              type:"error",
-              message:"删除失败！"
-            })
-          }
-        });
+        if(this.multipleSelection.length>0){
+          this.$http.delete(this.$url.baseUrl+'api/guoYang/rescue-team/v0.1/team',{data:this.multipleSelection}).then((res)=>{
+            if(res.status===200){
+              this.$message({
+                type:"success",
+                message:"删除成功！"
+              });
+              this.loading=true;
+              this.$refs.multipleTable.clearSelection();
+              this.search();
+            }else{
+              this.$message({
+                type:"error",
+                message:"删除失败！"
+              })
+            }
+          });
+        }else{
+          this.$message({
+            type:"error",
+            message:"请选择需要删除的数据！"
+          })
+        }
+
       },
       add(){
         this.dialogVisible=true;
@@ -413,24 +422,14 @@
                 _this.multipleSelection=[];
                 _this.search();
                 _this.$refs['ruleForm'].resetFields();
-                this.ruleForm={
-                  rtName:"",
-                  rtType:'常驻抢险队伍',
-                  teamMemberInfo:"",
-                  userId:"",
-                };
+                _this.dialogVisible=false;
               }else{
                 _this.$message({
                   type:"error",
                   message:msg1
                 });
                 _this.$refs['ruleForm'].resetFields();
-                this.ruleForm={
-                  rtName:"",
-                  rtType:'常驻抢险队伍',
-                  teamMemberInfo:"",
-                  userId:"",
-                };
+                _this.dialogVisible=false;
               }
             })
           } else {
